+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/notification/domain/contract/EdcNotificationContractService.java
@@ -17,6 +17,7 @@
  * under the License.
  *
  * SPDX-License-Identifier: Apache-2.0
+ *
  ********************************************************************************/
 package org.eclipse.tractusx.traceability.notification.domain.contract;